refactor(output): extract template tests to sibling file
